+++ src/parser.rs
@@ -499,14 +499,11 @@ mod tests {
         successful_test_case("let constructor : C k = C e;");
         successful_test_case("let pi_lambda : \\Pi0 a : b . c = \\lambda a . expr;");
         successful_test_case("let pat, pat2 : \\Pi0 _ : b . c = \\lambda _ . expr;");
+        successful_test_case("let function : Sum {C e} = split {C _ => e};");
     }
 
     #[test]
     fn no_reparse() {
-        successful_no_reparse("let sigma_type : \\Sigma x : x_type . y = x, y;");
-        successful_no_reparse("let pi_lambda : \\Pi a : b . c = \\lambda a . expr;");
-        successful_no_reparse("let pat, pat2 : \\Pi _ : b . c = \\lambda _ . expr;");
-        successful_no_reparse("let function : Sum {C e} = split {C _ => e};");
         successful_no_reparse("let function (x : a) : bla = rua;");
     }